requirements.txt
asgiref==3.8.0
Django==5.0.3
django-extensions==3.2.3
django-filter==24.1
djangorestframework==3.15.0
djangorestframework-jsonapi==6.1.0
inflection==0.5.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
